게시물 작성/수정/삭제 기능에 대해 인증 추가

+++ src/main/java/com/example/golfplatform/post/service/PostService.java
@@ -1,5 +1,6 @@
 package com.example.golfplatform.post.service;
 
+import com.example.golfplatform.exception.UnauthorizedAccessException;
 import com.example.golfplatform.post.domain.Post;
 import com.example.golfplatform.post.repository.PostRepository;
 import com.example.golfplatform.post.request.PostCreateRequest;
@@ -8,6 +9,7 @@ import com.example.golfplatform.post.response.PostDetailResponse;
 import com.example.golfplatform.post.response.PostListResponse;
 import com.example.golfplatform.user.domain.User;
 import com.example.golfplatform.user.repository.UserRepository;
+import java.nio.file.AccessDeniedException;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
@@ -31,9 +33,9 @@ public class PostService {
         return PostDetailResponse.from(post);
     }
 
-    public void createPost(PostCreateRequest request) {
-        User user = userRepository.findById(1L)
-            .orElseThrow(() -> new IllegalArgumentException(("사용자가 존재하지 않습니다.")));
+    public void createPost(PostCreateRequest request, Long userId) {
+        User user = userRepository.findById(userId)
+            .orElseThrow(() -> new RuntimeException(("유저를 찾을 수 없습니다.")));
         Post post = Post.builder()
             .user(user)
             .title(request.title())
@@ -43,15 +45,22 @@ public class PostService {
         postRepository.save(post);
     }
 
-    public void updatePost(Long id, PostUpdateRequest request) {
+    public void updatePost(Long id, PostUpdateRequest request, Long userId) {
         Post post = postRepository.findById(id)
             .orElseThrow(() -> new IllegalArgumentException("해당 게시글은 이미 존재하지 않습니다."));
+        if(!post.getUser().getId().equals(userId)) {
+            throw new UnauthorizedAccessException("수정 권한이 없습니다.");
+        }
         post.update(request.title(), request.content(), request.category());
     }
 
-    public void deletePost(Long id) {
+    public void deletePost(Long id, Long userId) {
         Post post = postRepository.findById(id)
             .orElseThrow(() -> new IllegalArgumentException("해당 게시글은 이미 존재하지 않습니다."));
+        if (!post.getUser().getId().equals(userId)) {
+            throw new UnauthorizedAccessException("삭제 권한이 없습니다.");
+        }
+
         postRepository.delete(post);
     }
